// Source/DocuEye.Web/client-app/src/api/docueye-api/models/element.ts
/* tslint:disable */
/* eslint-disable */
/**
 * DocuEye.Web
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface Element
 */
export interface Element {
    /**
     * 
     * @type {string}
     * @memberof Element
     */
    'id'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Element
     */
    'dslId'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Element
     */
    'name'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Element
     */
    'url'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Element
     */
    'group'?: string | null;
    /**
     * 
     * @type {Array<string>}
     * @memberof Element
     */
    'tags'?: Array<string> | null;
    /**
     * 
     * @type {string}
     * @memberof Element
     */
    'type'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Element
     */
    'location'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Element
     */
    'technology'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Element
     */
    'description'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Element
     */
    'parentId'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Element
     */
    'instanceOfId'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Element
     */
    'workspaceId'?: string | null;
    /**
     * 
     * @type {{ [key: string]: string; }}
     * @memberof Element
     */
    'properties'?: { [key: string]: string; } | null;
    /**
     * 
     * @type {string}
     * @memberof Element
     */
    'sourceCodeUrl'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Element
     */
    'ownerTeam'?: string | null;
}