// Source/DocuEye.Web/client-app/src/api/docueye-api/models/view-configuration.ts
/* tslint:disable */
/* eslint-disable */
/**
 * DocuEye.Web
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { ElementStyle } from './element-style';
// May contain unused imports in some cases
// @ts-ignore
import { RelationshipStyle } from './relationship-style';
// May contain unused imports in some cases
// @ts-ignore
import { Terminology } from './terminology';

/**
 * 
 * @export
 * @interface ViewConfiguration
 */
export interface ViewConfiguration {
    /**
     * 
     * @type {string}
     * @memberof ViewConfiguration
     */
    'id'?: string | null;
    /**
     * 
     * @type {Array<ElementStyle>}
     * @memberof ViewConfiguration
     */
    'elementStyles'?: Array<ElementStyle> | null;
    /**
     * 
     * @type {Array<RelationshipStyle>}
     * @memberof ViewConfiguration
     */
    'relationshipStyles'?: Array<RelationshipStyle> | null;
    /**
     * 
     * @type {Array<string>}
     * @memberof ViewConfiguration
     */
    'themes'?: Array<string> | null;
    /**
     * 
     * @type {Terminology}
     * @memberof ViewConfiguration
     */
    'terminology'?: Terminology;
    /**
     * 
     * @type {string}
     * @memberof ViewConfiguration
     */
    'groupSeparator'?: string | null;
}